requirements.txt
Flask==3.0.0
mysql-connector-python==8.2.0
pytest==7.4.3
Werkzeug==3.0.1
coverage==7.3.4
pytest-cov==4.1.0
flask-cors

